Move ProcedureShouldBeFunctionInspectionTests for QuickFixes into relevant files under RD.QuickFixes
