+++ RubberduckTests/RubberduckTests.csproj
@@ -108,6 +108,7 @@
     <Compile Include="Inspections\MemberNotOnInterfaceInspectionTests.cs" />
     <Compile Include="Inspections\StopKeywordInspectionTests.cs" />
     <Compile Include="Inspections\OptionBaseZeroInspectionTests.cs" />
+    <Compile Include="QuickFixes\ChangeProcedureToFunctionQuickFixTests.cs" />
     <Compile Include="QuickFixes\PassParameterByReferenceQuickFixTests.cs" />
     <Compile Include="Inspections\QuickFixProviderTests.cs" />
     <Compile Include="Inspections\SynchronizeAttributesQuickFixTests.cs" />
